+++ src/pages/AddReviewPage.tsx
@@ -1,51 +1,26 @@
 import { useState } from 'react';
-import api from '@/api/axios';
 import { ImageUploader } from '@/components/addreview/ImageUploader';
 import { TextInput } from '@/components/common/TextInput';
 import Button from '@/components/common/Button';
 import CafeSearchSection from '@/components/addreview/CafeSearchSection';
 import RatingSliderList from '@/components/addreview/RatingSliderList';
+import { useSubmitReview } from '@/hooks/useSubmitReview';
+
+const INITIAL_RATINGS: Record<string, number> = {
+  study: 3, outlet: 3, seat: 3, toilet: 3, wifi: 3, noise: 3,
+};
 
 const AddReviewPage = () => {
   const [cafeId] = useState(1);
   const [cafeName] = useState('스타벅스 서강대점');
   const [comment, setComment] = useState('');
   const [selectedFile, setSelectedFile] = useState<File | null>(null);
+  const [ratings, setRatings] = useState<Record<string, number>>(INITIAL_RATINGS);
 
-  const [ratings, setRatings] = useState<Record<string, number>>({
-    study: 3, outlet: 3, seat: 3, toilet: 3, wifi: 3, noise: 3,
-  });
+  const { submitReview } = useSubmitReview();
 
   const handleRatingChange = (key: string, value: number) => {
     setRatings((prev) => ({ ...prev, [key]: value }));
-  };
-
-  const handleSubmit = async () => {
-    const formData = new FormData();
-
-    const reviewData = {
-      cafeId,
-      shortReview: comment,
-      studyScore: ratings.study,
-      outletScore: ratings.outlet,
-      seatScore: ratings.seat,
-      toiletScore: ratings.toilet,
-      wifiScore: ratings.wifi,
-      noiseScore: ratings.noise,
-    };
-
-    formData.append('data', new Blob([JSON.stringify(reviewData)], { type: 'application/json' }));
-    if (selectedFile) formData.append('image', selectedFile);
-
-    try {
-      const response = await api.post('/api/reviews', formData, {
-        headers: { 'Content-Type': 'multipart/form-data' },
-      });
-      alert('성공적으로 저장되었습니다! 리뷰 ID: ' + response.data);
-    } catch (error) {
-      console.error(error);
-      alert('저장 중 오류가 발생했습니다.');
-    }
   };
 
   return (
@@ -65,7 +40,7 @@ const AddReviewPage = () => {
         <TextInput placeholder="한 줄 리뷰를 작성해주세요" value={comment} onChange={setComment} />
       </div>
 
-      <Button variant="brown4" size="full" onClick={handleSubmit}>
+      <Button variant="brown4" size="full" onClick={() => submitReview({ cafeId, comment, ratings, selectedFile })}>
         저장
       </Button>
     </div>
